Add project stack.yaml and cabal.project files
Add support for building 'natural4' with `cabal` as well as `stack`.
We add a `cabal.project` file that mimics the corresponding `stack.yaml`
file as closely as possible.
Note, the `cabal.project` file is maintained on a best-effort basis and
may sometimes fall behind the `stack.yaml`, which is the primary way of
building the project.
Haskell-Language-Server reportedly works more reliable with `cabal`,
which is why we are giving this a try.

Add singular, global `stack.yaml` file which is the single source of
truth for `extra-deps`, the `resolver` and globally applied `ghc-options`.
A single `stack.yaml` may also lead to an improved HLS experience, as
the language server might get confused by multiple `stack.yaml` files.
To keep the same IDE experience, we add a `hie.yaml` which outlines the
project structure to HLS.

Update the READMEs to reflect small changes in `stack install`
differences. Most notably, `stack run` now requires the name of the
executable to be unambiguous. For example, to run `natural4`, the
invocation is now:

  stack run natural4-exe -- ...

At last, we commit the generated .cabal files to the git repository, as
per the recommendation: https://www.fpcomplete.com/blog/storing-generated-cabal-files/