Add OpenAPI specification and swagger-ui
In order for LLMs to interact with the REST API (which is on the goals
of the so-called LAG project), we need to provide an openapi3 compatible
API.
We do this by depending on 'servant-openapi3', and add appropriate
instances for the servant REST API.

Due to some issues, it seems like it is impossible (or very difficult),
to make custom gpts invoke REST endpoints with JSON bodies... So, for
now we simply send all parameters via Query Parameters.
Until the next servant release, we have to explicitly name all
query parameters. See
https://github.com/haskell-servant/servant/pull/1604 for the PR that we
are interested in.