Don't hardcode the Effect stack
For code reuse, it is generally advisable to not group all potentially
required effects and make every function depend on it. That's basically
a monad transformer stack with extra steps.

We make sure that each function depends on only the effects it actually
needs. On top of being more iditomatic, it allows us to easily argue
which effects a function may use.