Introduce structured IR of MtExpr
The new IR (Intermediate Representation) introduces a Tree-like
structure for the program AST in favour of the unstructured list.
This avoids some additional work in the transpiler, as the transpiler no
longer needs to figure out what the program AST actually means, at the
cost of some additional complexity in the renamer. However, since the
complexity has been moved from one point to the other, we end with a
nice separation, keeping the transpiler simple and focusing on the
individual transpiler, and tucking complexity away into the Renamer
phase.

We also improve the Simala backend slightly by utilising the new program
AST.