Add support for Array and Object inputs
OIA input variables may consist arrays and objects.
So, we enhance `FnLiteral` to be essentially a `Aeson.Value` with native
support for `uncertain` and `unknown`.

A missing value (`null` or absent) is inferred to be `unknown`.
An empty object (`{}`) represents a `uncertain` value.

Only the simala backend can handle OIA values, such as arrays, objects
and `unknown`/`uncertain`.

We should now be able to handle all OIA batch requests, however, we may
ignore configurations, such as `reportSilent` and `decision-report`
style.